# backend/docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      data: {}
      message:
        description: 用于存储自定义消息
        type: string
    type: object
...
  model.DeviceAuthReq:
    properties:
      device_name:
        description: 设备名称(可选)
        maxLength: 255
        type: string
      device_number:
        description: 设备唯一标识
        maxLength: 255
        type: string
      product_key:
        description: 产品密钥(可选，用于产品关联)
        maxLength: 255
        type: string
      template_secret:
        description: 模板密钥
        maxLength: 255
        type: string
    required:
    - device_number
    - template_secret
    type: object
  model.DeviceAuthRes:
    properties:
      device_id:
        description: 设备ID
        type: string
      voucher:
        description: 设备凭证
        type: string
    type: object
  model.LoginReq:
    properties:
      email:
        description: 登录账号(输入邮箱或者手机号)
        example: user@example.com
        type: string
      password:
        description: 密码
        example: "123456"
        maxLength: 512
        minLength: 6
        type: string
      salt:
        description: 随机盐(如果在超管设置了前端RSA加密则需要上送)
        maxLength: 512
        type: string
    required:
    - email
    - password
    type: object
  model.LoginRsp:
    properties:
      expires_in:
        description: 过期时间(单位:秒)
        type: integer
      token:
        description: 登录凭证
        type: string
    type: object
...
host: localhost:9999
info:
  contact: {}
  description: AetherLink IoT API.
  title: AetherLink IoT API
  version: "1.0"
paths:
...
  /api/v1/device/auth:
    post:
      consumes:
      - application/json
      description: 实现一型一密认证机制，设备通过此接口获取凭证
      parameters:
      - description: 认证请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeviceAuthReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.DeviceAuthRes'
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 设备动态认证
      tags:
      - 设备认证
...
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: 认证令牌(Token)将在用户成功登录后生成并返回。客户端需要在后续所有需要认证的API请求中，将此令牌添加到HTTP请求头(Header)的'x-token'字段中。服务器将通过验证此令牌来确认用户身份并授权访问受保护资源。
      parameters:
      - description: 登录凭证
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/model.LoginRsp'
        "400":
          description: 错误响应
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户登录
      tags:
      - 用户认证
...
  /api/v1/plugin/device/config:
    post:
      consumes:
      - application/json
      description: Plugin-side device configuration lookup. Authenticate with the X-API-Key header.
      parameters:
      - in: body
        name: request
        schema:
          type: object
      responses: {}
      security:
      - PluginApiKeyAuth: []
      summary: Plugin device configuration
      tags:
      - Plugin OpenAPI
  /api/v1/plugin/devices:
    post:
      consumes:
      - application/json
      description: Plugin-side device list lookup by service identifier and device type. Authenticate with the X-API-Key header.
      parameters:
      - in: body
        name: request
        schema:
          type: object
      responses: {}
      security:
      - PluginApiKeyAuth: []
      summary: Plugin device list
      tags:
      - Plugin OpenAPI
  /api/v1/plugin/service/access:
    post:
      consumes:
      - application/json
      description: Plugin-side service access detail lookup. Authenticate with the X-API-Key header.
      parameters:
      - in: body
        name: request
        schema:
          type: object
      responses: {}
      security:
      - PluginApiKeyAuth: []
      summary: Plugin service access detail
      tags:
      - Plugin OpenAPI
  /api/v1/plugin/service/access/list:
    post:
      consumes:
      - application/json
      description: Plugin-side service access list. Authenticate with the X-API-Key header and send service_identifier in the request body.
      parameters:
      - in: body
        name: request
        schema:
          properties:
            service_identifier:
              type: string
          required:
          - service_identifier
          type: object
      responses: {}
      security:
      - PluginApiKeyAuth: []
      summary: Plugin service access list
      tags:
      - Plugin OpenAPI
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
  PluginApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
swagger: "2.0"